lambdas/requirements.txt
requests==2.25.1
boto3
PyJWT